pl/math: Refine ulp assessment for log10
Somehow the tests currently miss a lot of values which give
significantly worse error than the previous estimate. Attempted to
come up with a new absolute-worst-case estimate, however it is still
very much an estimate as values close to the worst-case appear right
across the range of positive doubles, with tiny and huge values being
a shade worse.